requirements.txt
requests~=2.28.0
jsons~=1.6.3
python-dotenv~=0.20.0